Allow navigating away from dive screen. Rehydrate dive state when opening dive screen again.
